Changed trials to store long

+++ app/src/main/java/xyz/kotlout/kotlout/model/experiment/trial/CountTrial.java
@@ -5,22 +5,22 @@ package xyz.kotlout.kotlout.model.experiment.trial;
  */
 public class CountTrial extends Trial {
 
-  private int result;
+  private long result;
 
   public CountTrial() {
 
   }
 
-  public CountTrial(int result, String experimenterId) {
+  public CountTrial(long result, String experimenterId) {
     super(experimenterId);
     this.result = result;
   }
 
-  public int getResult() {
+  public long getResult() {
     return result;
   }
 
-  public void setResult(int result) {
+  public void setResult(long result) {
     this.result = result;
   }
 }
